move Node text rendering to label provider
