+++ org.scala-ide.sdt.core/src/org/scalaide/ui/internal/editor/outline/ModelBuilder.scala
@@ -1,4 +1,17 @@
 package org.scalaide.ui.internal.editor.outline
+
+import scala.tools.nsc.Settings
+import scala.tools.nsc.interactive.CompilerControl
+import scala.tools.nsc.interactive.Global
+import scala.tools.nsc.reporters.ConsoleReporter
+import scala.reflect.internal.util.BatchSourceFile
+import scala.reflect.internal.util.SourceFile
+import scala.tools.nsc.interactive.Response
+import org.scalaide.core.compiler.IScalaPresentationCompiler
+import org.scalaide.logging.HasLogger
+import scala.collection.mutable.MutableList
+
+
 /**
  * ModelBuilder is the "heart" of Outline View. It maps scala source into a tree of nodes (model). Each node
  * represent a scala entity, such as Class, Trait, Method and so on.
@@ -15,229 +28,6 @@ package org.scalaide.ui.internal.editor.outline
  *
  * Using mutable model allows to update only nodes are modified.
  */
-import scala.tools.nsc.Settings
-import scala.tools.nsc.interactive.CompilerControl
-import scala.tools.nsc.interactive.Global
-import scala.tools.nsc.reporters.ConsoleReporter
-import scala.reflect.internal.util.BatchSourceFile
-import scala.reflect.internal.util.SourceFile
-import scala.tools.nsc.interactive.Response
-import org.scalaide.core.compiler.IScalaPresentationCompiler
-import org.scalaide.logging.HasLogger
-import scala.collection.mutable.MutableList
-
-trait HasReturnType {
-  private var rt: Option[String] = None
-  def returnType = rt
-  def returnType_=(rt: Option[String]) = {
-    this.rt = rt
-  }
-}
-trait HasModifiers {
-  import scala.reflect.internal.Flags._
-  private[outline] var flags = 0L
-  def setFlags(f: Long) = {
-    flags = f
-  }
-  def isPrivate = (flags & PRIVATE) != 0
-  def isProtected = (flags & PROTECTED) != 0
-  def isTrait = (flags & TRAIT) != 0
-  def isParam = (flags & PARAM) != 0
-  def isSynthetic = (flags & SYNTHETIC) != 0
-  def isImplicit = (flags & IMPLICIT) != 0
-}
-sealed abstract class NodeKey
-
-sealed abstract class Node(val isLeaf: Boolean, val parent: ContainerNode) extends HasModifiers {
-  def name: String
-  var start: Int = 0
-  var end: Int = 0
-  def displayName = name
-  def key: NodeKey
-  /**
-   * makes this node equal to source. Returns true if any changes are visible.
-   * For example, if user just added a new line, node position (start, end) is different now. But this change is not visible, so
-   * no updates should be generated.
-   *
-   * @param src a node to make equal to
-   * @return true if any changes are visible, false otherwise
-   */
-  def update(src: Node): Boolean = {
-    start = src.start
-    end = src.end
-    val f = flags
-    flags = src.flags
-    f != src.flags
-  }
-}
-
-sealed abstract class ContainerNode(parent: ContainerNode) extends Node(false, parent) {
-  protected var _children: Map[NodeKey, Node] = new scala.collection.immutable.ListMap[NodeKey, Node]()
-  def children = _children
-  def children_=(ch: Map[NodeKey, Node]) = _children = ch
-  def addChild(ch: Node) = {
-    last = Some(ch)
-    _children += (ch.key -> ch)
-  }
-  var last: Option[Node] = None
-}
-
-case class ClassKey(name: String) extends NodeKey
-case class ClassNode(name: String, override val parent: ContainerNode) extends ContainerNode(parent) {
-  override def key = ClassKey(name)
-}
-
-case class ObjectKey(name: String) extends NodeKey
-case class ObjectNode(name: String, override val parent: ContainerNode) extends ContainerNode(parent) {
-  override def key = ObjectKey(name)
-}
-
-case class TraitKey(name: String) extends NodeKey
-case class TraitNode(name: String, override val parent: ContainerNode) extends ContainerNode(parent) {
-  override def key = ObjectKey(name)
-}
-
-case class PackageKey(name: String) extends NodeKey
-case class PackageNode(name: String, override val parent: ContainerNode) extends Node(true, parent) {
-  override def key = PackageKey(name)
-}
-
-case class RootNode() extends ContainerNode(null) {
-  override def name = "ROOT"
-  override def key = null
-  /**
-   * iterates through whole tree and makes it equal to src.
-   * It returns two sets of nodes- with structural changes (some node where added or removed) and content changes.
-   *
-   * @param src new generated tree
-   * @return nodes with structural and content difference.
-   */
-  def updateAll(src: RootNode): (Iterable[Node], Iterable[Node]) = {
-    val toUpdate = new MutableList[Node]
-    val toRefresh = new MutableList[Node]
-    def visitContainer(cn: ContainerNode, cn1: ContainerNode): Unit = {
-      var children = collection.immutable.ListMap[NodeKey, Node]()
-      var needUpdate = false
-
-      cn1.children.foreach(p =>
-        cn.children.get(p._1) match {
-          case Some(n) =>
-            if (visitNode(n, p._2))
-              needUpdate = true
-            children += (p._1 -> n)
-          case None =>
-            needUpdate = true
-            children += (p._1 -> p._2)
-        })
-      if (children.size != cn.children.size)
-        needUpdate = true
-      if (needUpdate)
-        toUpdate += cn
-      cn.children = children
-    }
-
-    def visitNode(n: Node, n1: Node): Boolean = {
-      val b = n.update(n1)
-      if (b)
-        toRefresh += n
-      n match {
-        case cn: ContainerNode =>
-          n1 match {
-            case cn1: ContainerNode =>
-              visitContainer(cn, cn1)
-            case _ =>
-          }
-        case _ =>
-      }
-      b
-    }
-    visitContainer(this, src)
-    (toUpdate, toRefresh)
-  }
-}
-
-case class MethodKey(name: String, pList: List[List[String]] = List()) extends NodeKey
-case class MethodNode(name: String, override val parent: ContainerNode, val argTypes: List[List[String]] = List())
-    extends ContainerNode(parent) with HasModifiers with HasReturnType {
-  override def displayName = {
-    val sb = new StringBuilder
-    def renderArgList() = {
-      argTypes.foreach(list => {
-        sb.append("(")
-        list.foreach { s => sb.append(s); sb.append(", ") }
-        if (!list.isEmpty)
-          sb.setLength(sb.length - 2)
-        sb.append(")")
-      })
-    }
-    sb.append(name)
-    renderArgList
-    sb.append(returnType.map(": " + _).getOrElse(""))
-    sb.toString
-  }
-  override def key = MethodKey(name, argTypes)
-  override def update(n: Node): Boolean = {
-    val b = super.update(n)
-    n match {
-      case that: MethodNode =>
-        val rt = returnType
-        returnType = that.returnType
-        b || rt != returnType
-      case _ => b
-    }
-  }
-}
-
-case class ValNode(name: String, override val parent: ContainerNode, rt: Option[String])
-    extends Node(true, parent) with HasModifiers with HasReturnType {
-  returnType = rt
-  override def displayName = name + returnType.map(": " + _).getOrElse("")
-  override def key = MethodKey(name)
-  override def update(n: Node): Boolean = {
-    val b = super.update(n)
-    n match {
-      case that: ValNode =>
-        val rt = returnType
-        returnType = that.returnType
-        b || rt != returnType
-      case _ => b
-    }
-  }
-}
-
-case class VarNode(name: String, override val parent: ContainerNode, rt: Option[String])
-    extends Node(true, parent) with HasModifiers with HasReturnType {
-  returnType = rt
-  override def displayName = name + returnType.map(": " + _).getOrElse("")
-  override def key = MethodKey(name)
-  override def update(n: Node): Boolean = {
-    val b = super.update(n)
-    n match {
-      case that: VarNode =>
-        val rt = returnType
-        returnType = that.returnType
-        b || rt != returnType
-      case _ => b
-    }
-  }
-}
-case class TypeKey(name: String) extends NodeKey
-case class TypeNode(name: String, override val parent: ContainerNode)
-    extends Node(true, parent) with HasModifiers {
-  override def displayName = name
-  override def key = TypeKey(name)
-}
-
-case class ImportNode(name: String, override val parent: ContainerNode)
-    extends Node(true, parent) with HasModifiers {
-  override def key = ObjectKey(name)
-}
-
-case class ImportsNode(override val parent: ContainerNode) extends ContainerNode(parent) {
-  override def name = "import declarations"
-  override def key = ObjectKey(name)
-}
-
 object ModelBuilder  extends HasLogger{
   def buildTree(comp: IScalaPresentationCompiler, src: SourceFile): RootNode = {
     import comp._
